Authorization Finished

+++ mean-course/backend/routes/posts.js
@@ -30,39 +30,47 @@ const storage = multer.diskStorage({
 router.post("",
   checkAuth,
   multer({storage: storage}).single("image"), (req, res, next) => {
-  const url = req.protocol + '://' + req.get("host");
-  const post = new Post({
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
-  });
-  post.save().then(createdPost => {
-    res.status(201).json({
-      message: 'Post added sucessfully',
-      post: {
-        ...createdPost,
-        id: createdPost._id
-      }
+    const url = req.protocol + '://' + req.get("host");
+    const post = new Post({
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
+    });
+    post.save().then(createdPost => {
+      res.status(201).json({
+        message: 'Post added sucessfully',
+        post: {
+          ...createdPost,
+          id: createdPost._id
+        }
+      });
     });
   });
-});
 
-router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
-  let imagePath = req.body.imagePath;
-  if (req.file) {
-    const url = req.protocol + '://' + req.get("host");
-    imagePath = url + "/images/" + req.file.filename
-  }
-  const post = new Post({
-    _id: req.body.id,
-    title: req.body.title,
-    content: req.body.content,
-    imagePath: imagePath
+router.put("/:id",
+  checkAuth,
+  multer({storage: storage}).single("image"), (req, res, next) => {
+    let imagePath = req.body.imagePath;
+    if (req.file) {
+      const url = req.protocol + '://' + req.get("host");
+      imagePath = url + "/images/" + req.file.filename
+    }
+    const post = new Post({
+      _id: req.body.id,
+      title: req.body.title,
+      content: req.body.content,
+      imagePath: imagePath,
+      creator: req.userData.userId
+    });
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({message: "Not Authorized!"});
+      }
+    })
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({message: "Update successful!"})
-  })
-});
 
 router.get("", (req, res, next) => {
   const pageSize = +req.query.pagesize;
@@ -98,8 +106,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    res.status(200).json({message: 'Post deleted!'});
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: "Deletion successful!"});
+    } else {
+      res.status(401).json({message: "Not Authorized!"});
+    }
   });
 });
